Match titlebar colors to GTK theme.
CC: https://github.com/vinceliuice/Qogir-theme/blob/master/src/gtk-3.0/theme/gtk-dark.css#L10520
The issue with Aurorae is that it doesn't allow non-255 colors. Well, it
does, but it doesn't work (it will be transparent to the whole
background, not only to the title), so what I did was put a background
of #282a33 on GIMP on one layer, then a foreground of #e4ebf1 and put it
on 80% opacity, then grab the color using KColorChooser. Same for the
unfocused title, but that one used #e4ebf1 at 50% opacity. Cursed, I
know.